:bento: Chore: 즐겨찾기 수정 data class 정의

+++ app/src/main/java/com/example/snack4diet/api/updateFavoriteFood/UpdateFavoriteFoodDto.kt
@@ -2,7 +2,7 @@ package com.example.snack4diet.api.updateFavoriteFood
 
 data class UpdateFavoriteFoodDto(
     val baseNutrition: BaseNutrition,
-    val favoriteFoodId: Int,
+    val favoriteFoodId: Long,
     val name: String,
-    val userId: Int
+    val userId: Long
 )
